Work in progress: Refactor file structure, implement Harsh suggestions
