+++ app/components/NDVI/Info.js
@@ -2,22 +2,10 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 const Info = props => {
-  // const dummyDataPoints = {
-  //   "monthUid": 201907,
-  //   "month": 7,
-  //   "year": 2019,
-  //   "carbon": 3716,
-  //   "ndviAggregate": {
-  //     "min": -0.24662007507553332,
-  //     "max": 0.7517624439222388,
-  //     "avg": 0.3811577383449884
-  //   }
-  // }
-
   return (
     <div className="info-component">
       <ul>
-        <li>
+        {/* <li>
           <p>NDVI Result from April 1 - April 30,{props.year}</p>
           <p>
             Min: <b>{props.ndviAggregate.min}</b> Average:{' '}
@@ -25,7 +13,7 @@ const Info = props => {
               {props.ndviAggregate.max}
             </b>
           </p>
-        </li>
+        </li> */}
         <li>
           <button>?</button>
         </li>
@@ -33,10 +21,6 @@ const Info = props => {
     </div>
   );
 };
-
-// function convertDate() {
-//   return 'hello';
-// }
 
 export default Info;
