+++ web/widgets/NDVI/app/components/App.js
@@ -1,12 +1,23 @@
 import React, { Component } from 'react';
-import NDVIContainer from './NDVIContainer';
-// import PropTypes from 'prop-types';
+import NDVIContainer from '../../../../../app/components/NDVI/Index';
 
 export default class App extends Component {
   constructor(props) {
     super(props);
     this.state = {};
   }
+
+  // const dummyDataPoints = {
+  //   monthUid: 201907,
+  //   month: 7,
+  //   year: 2019,
+  //   carbon: 3716,
+  //   ndviAggregate: {
+  //     min: -0.24662007507553332,
+  //     max: 0.7517624439222388,
+  //     avg: 0.3811577383449884
+  //   }
+  // };
 
   render() {
     return (
